Restructure tests directory to match repo src code
