+++ tests/src/pages/Home/Home.test.tsx
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import Home from '../../src/pages/Home/Home';
+import Home from '../../../../src/pages/Home/Home';
 
 describe('Home', () => {
   it('should render the page correctly', () => {
